DB is fixed, now ADD Project and LIST Projects is working correctly
There were a couple problems I solved.  First off the reason why append
wasn't working was we were adding the script at the top and the ccs
selector wasn't finding the list.  Therefore it was adding to nowhere. I
solved that by including the script main.js at the bottom of the page.
Also you were closing the ResultSet from the DB before you read it and
that was releasing the memory and the results.