Python script successfully running and taking an asset as an arguement to the script
Trevor is working on a few bugs for the style, but it is functional at
this point.  Just add a python script that takes an arguement and it can
also output.  All of the output will be displayed as alerts, and you can
do whatever you would like with the selected arguement.  A alert is
called on script exit which will be useful in the future.