ScriptExecution Page finished.  Both Required and Optional Arguments
The page requires you to fill in all the required assets, but doesn't
care whether or not you fill the optional.  Trevor is going to work on
limiting the assets to 3 and overflow:scroll for anything >3.  The
Script status is working, but the only problem is that it will not
update until the script has finished executing.